+++ pkg/backup/restore_remote.go
@@ -2,7 +2,10 @@ package backup
 
 func (b *Backuper) RestoreFromRemote(backupName, tablePattern string, databaseMapping, partitions []string, schemaOnly, dataOnly, dropTable, ignoreDependencies, rbacOnly, configsOnly, resume bool, commandId int) error {
 	if err := b.Download(backupName, tablePattern, partitions, schemaOnly, resume, commandId); err != nil {
-		return err
+		// https://github.com/AlexAkulov/clickhouse-backup/issues/625
+		if err != ErrBackupIsAlreadyExists {
+			return err
+		}
 	}
 	return b.Restore(backupName, tablePattern, databaseMapping, partitions, schemaOnly, dataOnly, dropTable, ignoreDependencies, rbacOnly, configsOnly, commandId)
 }